requirements.txt
asgiref==3.2.3
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.0.9
defusedxml==0.6.0
dj-database-url==0.4.2
Django==1.11.4
django-allauth==0.32.0
django-angular==0.7.16
django-extensions==1.6.1
django_range_slider
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
django-widget-tweaks==1.4.1
google-api-python-client==1.5.0
gunicorn==19.6.0
httplib2==0.9.2
idna==2.8
ipdb==0.9.0
Markdown==2.6.9
oauth2client==2.0.1
oauthlib==3.1.0
path.py==8.1.2
pexpect==4.0.1
phonenumbers==8.11.3
pickleshare==0.6
Pillow==7.0.0
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==3.3
simplegeneric==0.8.1
simplejson==3.8.2
social-auth-app-django==1.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
traitlets==4.2.1
uritemplate==0.6
urllib3==1.25.8
psycopg2
django-multifilefield
